refactor(tower): namespace the middleware under http::server
The crate exposed a single flat set of names (HTTPLayer, HTTPLayerBuilder,
the route and attribute extractors) that all implicitly meant "HTTP
server". That leaves no room for the other protocols and directions the
crate is expected to cover, and readers cannot tell from a name whether
it applies to inbound or outbound traffic.

Group the middleware into `http::server`, move the pluggable extractors
into `http::extractors`, and pull the shared attribute helpers into a
private `common` module, so each name states which protocol and which
side of the connection it instruments. This is a rename and file move
only; the emitted spans, metrics, and attributes are unchanged.
