+++ opentelemetry-instrumentation-tower/Cargo.toml
@@ -35,7 +35,7 @@ tower = { version = "0.5", features = ["util"] }
 tower-test = { version = "0.4" }
 
 [[bench]]
-name = "middleware"
+name = "http_server"
 harness = false
 
 [lints]
