+++ opentelemetry-instrumentation-tower/src/common/mod.rs
@@ -1,0 +1,3 @@
+//! Building blocks shared by the HTTP layers.
+
+pub(crate) mod attributes;
