+++ opentelemetry-instrumentation-tower/src/lib.rs
@@ -4,6 +4,13 @@
 //! The middleware produces both metrics and distributed tracing for incoming HTTP
 //! requests, following the OpenTelemetry [HTTP semantic conventions].
 //!
+//! # Layout
+//!
+//! - [`http::server`] — instruments incoming requests (`SpanKind::Server`) and
+//!   extracts the trace context from request headers.
+//! - [`http::extractors`] — pluggable route and attribute extractors used by the
+//!   server layer.
+//!
 //! # Metrics
 //!
 //! - `http.server.request.duration` — duration of HTTP server requests.
@@ -20,11 +27,11 @@
 //! # Quick start
 //!
 //! With the default `axum` feature enabled, applying the middleware is a single
-//! [`HTTPLayer::new`] call:
+//! [`http::server::Layer::new`] call:
 //!
 //! ```ignore
 //! use axum::{routing::get, Router};
-//! use opentelemetry_instrumentation_tower::HTTPLayer;
+//! use opentelemetry_instrumentation_tower::http;
 //!
 //! # async fn root() -> &'static str { "hello" }
 //! # fn run() {
@@ -32,7 +39,7 @@
 //!     .route("/", get(root))
 //!     // Apply the OTel layer *after* the routes so that
 //!     // `AxumMatchedPathExtractor` can read the matched route template.
-//!     .layer(HTTPLayer::new());
+//!     .layer(http::server::Layer::new());
 //! # let _ = app;
 //! # }
 //! ```
@@ -42,23 +49,28 @@
 //!
 //! # Customization
 //!
-//! Use [`HTTPLayerBuilder`] to plug in custom extractors:
+//! Use [`http::server::LayerBuilder`] to plug in custom extractors:
 //!
-//! - [`RouteExtractor`] decides how the `http.route` attribute (and span name) is
-//!   produced. Built-in choices: [`NoRouteExtractor`], [`PathExtractor`],
-//!   [`AxumMatchedPathExtractor`] (requires the `axum` feature), or
-//!   [`FnRouteExtractor`].
-//! - [`RequestAttributeExtractor`] / [`ResponseAttributeExtractor`] let you attach
-//!   additional attributes to spans and metrics. The default is [`NoOpExtractor`].
+//! - [`http::extractors::RouteExtractor`] decides how the `http.route` attribute
+//!   (and span name) is produced. Built-in choices:
+//!   [`http::extractors::NoRouteExtractor`], [`http::extractors::PathExtractor`],
+//!   [`http::extractors::AxumMatchedPathExtractor`] (requires the `axum` feature),
+//!   or [`http::extractors::FnRouteExtractor`].
+//! - [`http::extractors::RequestAttributeExtractor`] /
+//!   [`http::extractors::ResponseAttributeExtractor`] let you attach additional
+//!   attributes to spans and metrics. The default is
+//!   [`http::extractors::NoOpExtractor`].
 //!
-//! See [`RouteExtractor`] for cardinality guidance — picking the wrong extractor
-//! can blow up the cardinality of your metrics.
+//! See [`http::extractors::RouteExtractor`] for cardinality guidance — picking the
+//! wrong extractor can blow up the cardinality of your metrics.
 //!
 //! # Cargo features
 //!
-//! - `axum` *(default-off)* — enables [`AxumMatchedPathExtractor`] and makes it the
-//!   default route extractor. Without this feature the default extractor is
-//!   [`NoRouteExtractor`] (method-only span names, no `http.route` attribute).
+//! - `axum` *(default-off)* — enables
+//!   [`http::extractors::AxumMatchedPathExtractor`] and makes it the default route
+//!   extractor. Without this feature the default extractor is
+//!   [`http::extractors::NoRouteExtractor`] (method-only span names, no
+//!   `http.route` attribute).
 //!
 //! # Examples
 //!
@@ -72,310 +84,15 @@
 //! [`MeterProvider`]: opentelemetry::metrics::MeterProvider
 //! [`examples/`]: https://github.com/open-telemetry/opentelemetry-rust-contrib/tree/main/opentelemetry-instrumentation-tower/examples
 
-use std::borrow::Cow;
-use std::future::Future;
-use std::pin::Pin;
-use std::string::String;
-use std::sync::Arc;
-use std::task::{Context, Poll};
-use std::time::Instant;
-use std::{fmt, result};
+use std::fmt;
 
-#[cfg(feature = "axum")]
-use axum::extract::MatchedPath;
-use opentelemetry::global::{self, BoxedTracer};
-use opentelemetry::metrics::Meter;
-use opentelemetry::metrics::{Histogram, UpDownCounter};
-use opentelemetry::trace::{SpanKind, Status, TraceContextExt, Tracer};
-use opentelemetry::Context as OtelContext;
-use opentelemetry::KeyValue;
-use opentelemetry_http::HeaderExtractor;
-use opentelemetry_semantic_conventions as semconv;
-use pin_project_lite::pin_project;
-use tower_layer::Layer;
-use tower_service::Service;
+mod common;
+pub mod http;
 
-const HTTP_SERVER_DURATION_METRIC: &str = semconv::metric::HTTP_SERVER_REQUEST_DURATION;
-const HTTP_SERVER_DURATION_UNIT: &str = "s";
+/// Instrumentation scope name reported on emitted spans and metrics.
+pub(crate) const INSTRUMENTATION_NAME: &str = "opentelemetry-instrumentation-tower";
 
-const OTEL_DEFAULT_HTTP_SERVER_DURATION_BOUNDS: [f64; 14] = [
-    0.005, 0.01, 0.025, 0.05, 0.075, 0.1, 0.25, 0.5, 0.75, 1.0, 2.5, 5.0, 7.5, 10.0,
-];
-
-const HTTP_SERVER_ACTIVE_REQUESTS_METRIC: &str = semconv::metric::HTTP_SERVER_ACTIVE_REQUESTS;
-const HTTP_SERVER_ACTIVE_REQUESTS_UNIT: &str = "{request}";
-
-const HTTP_SERVER_REQUEST_BODY_SIZE_METRIC: &str = semconv::metric::HTTP_SERVER_REQUEST_BODY_SIZE;
-const HTTP_SERVER_REQUEST_BODY_SIZE_UNIT: &str = "By";
-
-const HTTP_SERVER_RESPONSE_BODY_SIZE_METRIC: &str = semconv::metric::HTTP_SERVER_RESPONSE_BODY_SIZE;
-const HTTP_SERVER_RESPONSE_BODY_SIZE_UNIT: &str = "By";
-
-const NETWORK_PROTOCOL_NAME_LABEL: &str = semconv::attribute::NETWORK_PROTOCOL_NAME;
-const NETWORK_PROTOCOL_VERSION_LABEL: &str = semconv::attribute::NETWORK_PROTOCOL_VERSION;
-const URL_SCHEME_LABEL: &str = semconv::attribute::URL_SCHEME;
-
-const HTTP_REQUEST_METHOD_LABEL: &str = semconv::attribute::HTTP_REQUEST_METHOD;
-const HTTP_ROUTE_LABEL: &str = semconv::attribute::HTTP_ROUTE;
-const HTTP_RESPONSE_STATUS_CODE_LABEL: &str = semconv::attribute::HTTP_RESPONSE_STATUS_CODE;
-
-/// Trait for extracting the route/target from HTTP requests.
-///
-/// Implementations return an optional route string. When present, this is used for:
-/// - Span names: `"{method} {route}"` (e.g., `"GET /users/:id"`)
-/// - The `http.route` metric attribute
-///
-/// When `None` is returned, span names use method-only (e.g., `"GET"`) and no
-/// `http.route` attribute is set.
-///
-/// # Cardinality Considerations
-///
-/// The route should be a low-cardinality value (e.g., a route template like `/users/:id`)
-/// rather than the actual path (e.g., `/users/123`). High-cardinality routes can overwhelm
-/// the OpenTelemetry SDK's built-in cardinality limits and downstream backends.
-pub trait RouteExtractor<B>: Clone + Send + Sync + 'static {
-    /// Extracts the route from the request, if available.
-    ///
-    /// Returns `None` to use method-only span names and skip the `http.route` attribute.
-    fn extract_route(&self, req: &http::Request<B>) -> Option<String>;
-}
-
-/// Route extractor that returns no route (method-only span names).
-///
-/// This is the safest option as it avoids cardinality explosion from dynamic
-/// path segments. Span names will be just the HTTP method (e.g., `"GET"`).
-///
-/// The `http.route` attribute will not be set when using this extractor.
-#[derive(Clone, Default)]
-pub struct NoRouteExtractor;
-
-impl<B> RouteExtractor<B> for NoRouteExtractor {
-    fn extract_route(&self, _req: &http::Request<B>) -> Option<String> {
-        None
-    }
-}
-
-/// Route extractor that uses Axum's `MatchedPath` for low-cardinality routes.
-///
-/// This extractor uses the route template (e.g., `/users/:id`) instead of the actual
-/// path (e.g., `/users/123`), providing low-cardinality span names and route attributes
-/// that are safe for production use.
-///
-/// # When `MatchedPath` is unavailable
-///
-/// Returns `None` (falling back to method-only span names) when `MatchedPath` is not
-/// present in the request extensions. This can happen when:
-///
-/// - The OpenTelemetry layer is placed *before* Axum's router in the middleware stack.
-///   The layer must be placed *after* the router to access route information.
-/// - The request does not match any defined route (404 responses).
-/// - Using Axum's `fallback` handler, which does not set `MatchedPath`.
-///
-/// For correct route extraction, ensure the middleware order is:
-///
-/// ```ignore
-/// let app = Router::new()
-///     .route("/users/:id", get(handler))
-///     .layer(otel_layer);  // Layer applied after routes
-/// ```
-///
-/// See the [Axum documentation on middleware ordering](https://docs.rs/axum/latest/axum/middleware/index.html#ordering)
-/// for more details.
-#[cfg(feature = "axum")]
-#[derive(Clone, Default)]
-pub struct AxumMatchedPathExtractor;
-
-#[cfg(feature = "axum")]
-impl<B> RouteExtractor<B> for AxumMatchedPathExtractor {
-    fn extract_route(&self, req: &http::Request<B>) -> Option<String> {
-        req.extensions()
-            .get::<MatchedPath>()
-            .map(|matched_path| matched_path.as_str().to_owned())
-    }
-}
-
-/// Route extractor that uses the URL path (without query parameters).
-///
-/// # Warning: Cardinality
-///
-/// Using this extractor can cause **high cardinality** issues if your routes contain
-/// dynamic path segments (e.g., `/users/{id}`, `/orders/{order_id}/items/{item_id}`).
-/// Each unique path will create a unique span name and route attribute, potentially
-/// overwhelming your tracing and metrics backends with millions of unique series.
-///
-/// **Only use this if**:
-/// - Your routes are static (no path parameters)
-/// - You understand and accept the cardinality implications
-///
-/// Consider using a custom [`FnRouteExtractor`] with path normalization instead.
-#[derive(Clone, Default)]
-pub struct PathExtractor;
-
-impl<B> RouteExtractor<B> for PathExtractor {
-    fn extract_route(&self, req: &http::Request<B>) -> Option<String> {
-        Some(req.uri().path().to_owned())
-    }
-}
-
-/// A function-based route extractor.
-#[derive(Clone)]
-pub struct FnRouteExtractor<F> {
-    extractor: F,
-}
-
-impl<F> FnRouteExtractor<F> {
-    pub fn new(extractor: F) -> Self {
-        Self { extractor }
-    }
-}
-
-impl<F, B> RouteExtractor<B> for FnRouteExtractor<F>
-where
-    F: Fn(&http::Request<B>) -> Option<String> + Clone + Send + Sync + 'static,
-{
-    fn extract_route(&self, req: &http::Request<B>) -> Option<String> {
-        (self.extractor)(req)
-    }
-}
-
-#[cfg(feature = "axum")]
-type DefaultRouteExtractor = AxumMatchedPathExtractor;
-
-#[cfg(not(feature = "axum"))]
-type DefaultRouteExtractor = NoRouteExtractor;
-
-/// Trait for extracting custom attributes from HTTP requests
-pub trait RequestAttributeExtractor<B>: Clone + Send + Sync + 'static {
-    fn extract_attributes(&self, req: &http::Request<B>) -> Vec<KeyValue>;
-}
-
-/// Trait for extracting custom attributes from HTTP responses
-pub trait ResponseAttributeExtractor<B>: Clone + Send + Sync + 'static {
-    fn extract_attributes(&self, res: &http::Response<B>) -> Vec<KeyValue>;
-}
-
-/// Default implementation that extracts no attributes
-#[derive(Clone)]
-pub struct NoOpExtractor;
-
-impl<B> RequestAttributeExtractor<B> for NoOpExtractor {
-    fn extract_attributes(&self, _req: &http::Request<B>) -> Vec<KeyValue> {
-        vec![]
-    }
-}
-
-impl<B> ResponseAttributeExtractor<B> for NoOpExtractor {
-    fn extract_attributes(&self, _res: &http::Response<B>) -> Vec<KeyValue> {
-        vec![]
-    }
-}
-
-/// A function-based request attribute extractor
-#[derive(Clone)]
-pub struct FnRequestExtractor<F> {
-    extractor: F,
-}
-
-impl<F> FnRequestExtractor<F> {
-    pub fn new(extractor: F) -> Self {
-        Self { extractor }
-    }
-}
-
-impl<F, B> RequestAttributeExtractor<B> for FnRequestExtractor<F>
-where
-    F: Fn(&http::Request<B>) -> Vec<KeyValue> + Clone + Send + Sync + 'static,
-{
-    fn extract_attributes(&self, req: &http::Request<B>) -> Vec<KeyValue> {
-        (self.extractor)(req)
-    }
-}
-
-/// A function-based response attribute extractor
-#[derive(Clone)]
-pub struct FnResponseExtractor<F> {
-    extractor: F,
-}
-
-impl<F> FnResponseExtractor<F> {
-    pub fn new(extractor: F) -> Self {
-        Self { extractor }
-    }
-}
-
-impl<F, B> ResponseAttributeExtractor<B> for FnResponseExtractor<F>
-where
-    F: Fn(&http::Response<B>) -> Vec<KeyValue> + Clone + Send + Sync + 'static,
-{
-    fn extract_attributes(&self, res: &http::Response<B>) -> Vec<KeyValue> {
-        (self.extractor)(res)
-    }
-}
-
-/// State scoped to the entire middleware Layer.
-struct HTTPLayerState {
-    pub server_request_duration: Histogram<f64>,
-    pub server_active_requests: UpDownCounter<i64>,
-    pub server_request_body_size: Histogram<u64>,
-    pub server_response_body_size: Histogram<u64>,
-}
-
-#[derive(Clone)]
-/// [`Service`] used by [`HTTPLayer`]
-pub struct HTTPService<
-    S,
-    RouteExt = DefaultRouteExtractor,
-    ReqExt = NoOpExtractor,
-    ResExt = NoOpExtractor,
-> {
-    pub(crate) state: Arc<HTTPLayerState>,
-    route_extractor: RouteExt,
-    request_extractor: ReqExt,
-    response_extractor: ResExt,
-    inner_service: S,
-    tracer: Arc<BoxedTracer>,
-}
-
-#[derive(Clone)]
-/// [`Layer`] which applies the OTEL HTTP server metrics and tracing middleware
-pub struct HTTPLayer<
-    RouteExt = DefaultRouteExtractor,
-    ReqExt = NoOpExtractor,
-    ResExt = NoOpExtractor,
-> {
-    state: Arc<HTTPLayerState>,
-    route_extractor: RouteExt,
-    request_extractor: ReqExt,
-    response_extractor: ResExt,
-    tracer: Arc<BoxedTracer>,
-}
-
-impl HTTPLayer {
-    /// Create a new HTTP layer with default configuration using global providers
-    pub fn new() -> Self {
-        HTTPLayerBuilder::builder().build().unwrap()
-    }
-}
-
-impl Default for HTTPLayer {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-pub struct HTTPLayerBuilder<
-    RouteExt = DefaultRouteExtractor,
-    ReqExt = NoOpExtractor,
-    ResExt = NoOpExtractor,
-> {
-    meter: Option<Meter>,
-    req_dur_bounds: Option<Vec<f64>>,
-    route_extractor: RouteExt,
-    request_extractor: ReqExt,
-    response_extractor: ResExt,
-}
-
-/// Error typedef to implement `std::error::Error` for `opentelemetry_instrumentation_tower`
+/// Error type for `opentelemetry_instrumentation_tower`.
 pub struct Error {
     #[allow(dead_code)]
     inner: ErrorKind,
@@ -392,8 +109,8 @@ impl std::fmt::Display for Error {
 
 impl std::error::Error for Error {}
 
-/// `Result` typedef to use with the `opentelemetry_instrumentation_tower::Error` type
-pub type Result<T> = result::Result<T, Error>;
+/// `Result` typedef to use with the [`Error`] type.
+pub type Result<T> = std::result::Result<T, Error>;
 
 enum ErrorKind {
     #[allow(dead_code)]
@@ -408,1160 +125,5 @@ impl fmt::Debug for Error {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         f.debug_tuple("opentelemetry_instrumentation_tower::Error")
             .finish()
-    }
-}
-
-impl HTTPLayerBuilder {
-    pub fn builder() -> Self {
-        HTTPLayerBuilder {
-            meter: None,
-            req_dur_bounds: Some(Vec::from(OTEL_DEFAULT_HTTP_SERVER_DURATION_BOUNDS)),
-            route_extractor: DefaultRouteExtractor::default(),
-            request_extractor: NoOpExtractor,
-            response_extractor: NoOpExtractor,
-        }
-    }
-}
-
-impl<RouteExt, ReqExt, ResExt> HTTPLayerBuilder<RouteExt, ReqExt, ResExt> {
-    /// Set a custom route extractor.
-    ///
-    /// The route extractor determines how the route is extracted from requests.
-    /// The extracted route is used for:
-    /// - Span names: `"{method} {route}"` or just `"{method}"` if no route
-    /// - The `http.route` metric attribute
-    ///
-    /// See [`RouteExtractor`] for details on implementing custom extractors.
-    ///
-    /// # Example
-    ///
-    /// ```ignore
-    /// let layer = HTTPLayerBuilder::builder()
-    ///     .with_route_extractor(PathExtractor)
-    ///     .build()
-    ///     .unwrap();
-    /// ```
-    pub fn with_route_extractor<NewRoute>(
-        self,
-        extractor: NewRoute,
-    ) -> HTTPLayerBuilder<NewRoute, ReqExt, ResExt> {
-        HTTPLayerBuilder {
-            meter: self.meter,
-            req_dur_bounds: self.req_dur_bounds,
-            route_extractor: extractor,
-            request_extractor: self.request_extractor,
-            response_extractor: self.response_extractor,
-        }
-    }
-
-    /// Convenience method to set a function-based route extractor.
-    ///
-    /// # Example
-    ///
-    /// ```ignore
-    /// let layer = HTTPLayerBuilder::builder()
-    ///     .with_route_extractor_fn(|req: &http::Request<_>| {
-    ///         // Return Some(route) or None for method-only
-    ///         Some(req.uri().path().to_owned())
-    ///     })
-    ///     .build()
-    ///     .unwrap();
-    /// ```
-    pub fn with_route_extractor_fn<F, B>(
-        self,
-        f: F,
-    ) -> HTTPLayerBuilder<FnRouteExtractor<F>, ReqExt, ResExt>
-    where
-        F: Fn(&http::Request<B>) -> Option<String> + Clone + Send + Sync + 'static,
-    {
-        self.with_route_extractor(FnRouteExtractor::new(f))
-    }
-
-    /// Set a request attribute extractor
-    pub fn with_request_extractor<NewReqExt, B>(
-        self,
-        extractor: NewReqExt,
-    ) -> HTTPLayerBuilder<RouteExt, NewReqExt, ResExt>
-    where
-        NewReqExt: RequestAttributeExtractor<B>,
-    {
-        HTTPLayerBuilder {
-            meter: self.meter,
-            req_dur_bounds: self.req_dur_bounds,
-            route_extractor: self.route_extractor,
-            request_extractor: extractor,
-            response_extractor: self.response_extractor,
-        }
-    }
-
-    /// Set a response attribute extractor
-    pub fn with_response_extractor<NewResExt, B>(
-        self,
-        extractor: NewResExt,
-    ) -> HTTPLayerBuilder<RouteExt, ReqExt, NewResExt>
-    where
-        NewResExt: ResponseAttributeExtractor<B>,
-    {
-        HTTPLayerBuilder {
-            meter: self.meter,
-            req_dur_bounds: self.req_dur_bounds,
-            route_extractor: self.route_extractor,
-            request_extractor: self.request_extractor,
-            response_extractor: extractor,
-        }
-    }
-
-    /// Convenience method to set a function-based request extractor
-    pub fn with_request_extractor_fn<F, B>(
-        self,
-        f: F,
-    ) -> HTTPLayerBuilder<RouteExt, FnRequestExtractor<F>, ResExt>
-    where
-        F: Fn(&http::Request<B>) -> Vec<KeyValue> + Clone + Send + Sync + 'static,
-    {
-        self.with_request_extractor(FnRequestExtractor::new(f))
-    }
-
-    /// Convenience method to set a function-based response extractor
-    pub fn with_response_extractor_fn<F, B>(
-        self,
-        f: F,
-    ) -> HTTPLayerBuilder<RouteExt, ReqExt, FnResponseExtractor<F>>
-    where
-        F: Fn(&http::Response<B>) -> Vec<KeyValue> + Clone + Send + Sync + 'static,
-    {
-        self.with_response_extractor(FnResponseExtractor::new(f))
-    }
-
-    pub fn build(self) -> Result<HTTPLayer<RouteExt, ReqExt, ResExt>> {
-        let req_dur_bounds = self
-            .req_dur_bounds
-            .unwrap_or_else(|| Vec::from(OTEL_DEFAULT_HTTP_SERVER_DURATION_BOUNDS));
-
-        let tracer = Arc::new(global::tracer("opentelemetry-instrumentation-tower"));
-
-        let meter: Meter = self
-            .meter
-            .unwrap_or_else(|| global::meter("opentelemetry-instrumentation-tower"));
-
-        Ok(HTTPLayer {
-            state: Arc::from(Self::make_state(meter, req_dur_bounds)),
-            route_extractor: self.route_extractor,
-            request_extractor: self.request_extractor,
-            response_extractor: self.response_extractor,
-            tracer,
-        })
-    }
-
-    /// Override the meter used for metrics collection.
-    ///
-    /// This method exists primarily for testing purposes, allowing tests to inject
-    /// a custom meter (e.g., backed by an in-memory exporter) without relying on
-    /// global state. Using global providers in tests can cause interference between
-    /// concurrent tests.
-    ///
-    /// In production, the default behavior of using the global meter provider
-    /// (via `opentelemetry::global::meter()`) is recommended.
-    #[cfg(test)]
-    fn with_meter(mut self, meter: Meter) -> Self {
-        self.meter = Some(meter);
-        self
-    }
-
-    fn make_state(meter: Meter, req_dur_bounds: Vec<f64>) -> HTTPLayerState {
-        HTTPLayerState {
-            server_request_duration: meter
-                .f64_histogram(Cow::from(HTTP_SERVER_DURATION_METRIC))
-                .with_description("Duration of HTTP server requests.")
-                .with_unit(Cow::from(HTTP_SERVER_DURATION_UNIT))
-                .with_boundaries(req_dur_bounds)
-                .build(),
-            server_active_requests: meter
-                .i64_up_down_counter(Cow::from(HTTP_SERVER_ACTIVE_REQUESTS_METRIC))
-                .with_description("Number of active HTTP server requests.")
-                .with_unit(Cow::from(HTTP_SERVER_ACTIVE_REQUESTS_UNIT))
-                .build(),
-            server_request_body_size: meter
-                .u64_histogram(HTTP_SERVER_REQUEST_BODY_SIZE_METRIC)
-                .with_description("Size of HTTP server request bodies.")
-                .with_unit(HTTP_SERVER_REQUEST_BODY_SIZE_UNIT)
-                .build(),
-            server_response_body_size: meter
-                .u64_histogram(HTTP_SERVER_RESPONSE_BODY_SIZE_METRIC)
-                .with_description("Size of HTTP server response bodies.")
-                .with_unit(HTTP_SERVER_RESPONSE_BODY_SIZE_UNIT)
-                .build(),
-        }
-    }
-}
-
-impl<S, RouteExt, ReqExt, ResExt> Layer<S> for HTTPLayer<RouteExt, ReqExt, ResExt>
-where
-    RouteExt: Clone,
-    ReqExt: Clone,
-    ResExt: Clone,
-{
-    type Service = HTTPService<S, RouteExt, ReqExt, ResExt>;
-
-    fn layer(&self, service: S) -> Self::Service {
-        HTTPService {
-            state: self.state.clone(),
-            route_extractor: self.route_extractor.clone(),
-            request_extractor: self.request_extractor.clone(),
-            response_extractor: self.response_extractor.clone(),
-            inner_service: service,
-            tracer: self.tracer.clone(),
-        }
-    }
-}
-
-/// Request data extracted before the inner service call.
-/// This data is needed for metrics and span finalization after the response is received.
-struct RequestData {
-    // fields for the metric values
-    // https://opentelemetry.io/docs/specs/semconv/http/http-metrics/#metric-httpserverrequestduration
-    duration_start: Instant,
-    // https://opentelemetry.io/docs/specs/semconv/http/http-metrics/#metric-httpserverrequestbodysize
-    req_body_size: Option<u64>,
-
-    // fields for metric labels
-    protocol_name_kv: KeyValue,
-    protocol_version_kv: KeyValue,
-    url_scheme_kv: KeyValue,
-    method_kv: KeyValue,
-    route_kv_opt: Option<KeyValue>,
-
-    // Custom attributes from request
-    custom_request_attributes: Vec<KeyValue>,
-}
-
-/// Maps common HTTP methods to a `&'static str` so the resulting `KeyValue`
-/// stores the method as a static string (no heap allocation, allocation-free
-/// `KeyValue::clone()`). Returns `None` for custom/extension methods, which
-/// fall back to an owned `String`.
-#[inline]
-fn method_as_static(m: &http::Method) -> Option<&'static str> {
-    match *m {
-        http::Method::GET => Some("GET"),
-        http::Method::POST => Some("POST"),
-        http::Method::PUT => Some("PUT"),
-        http::Method::DELETE => Some("DELETE"),
-        http::Method::HEAD => Some("HEAD"),
-        http::Method::OPTIONS => Some("OPTIONS"),
-        http::Method::PATCH => Some("PATCH"),
-        http::Method::CONNECT => Some("CONNECT"),
-        http::Method::TRACE => Some("TRACE"),
-        _ => None,
-    }
-}
-
-struct RequestFinalization<ResExt> {
-    request_data: RequestData,
-    layer_state: Arc<HTTPLayerState>,
-    response_extractor: ResExt,
-}
-
-pin_project! {
-    /// Future type returned by [`HTTPService`].
-    ///
-    /// This is a concrete future that avoids heap allocation by embedding the
-    /// inner service future directly, rather than using `Pin<Box<dyn Future>>`.
-    pub struct ResponseFuture<F, ResExt> {
-        #[pin]
-        inner: F,
-        otel_cx: OtelContext,
-        finalization: Option<RequestFinalization<ResExt>>,
-    }
-}
-
-impl<F, ResBody, E, ResExt> Future for ResponseFuture<F, ResExt>
-where
-    F: Future<Output = result::Result<http::Response<ResBody>, E>>,
-    E: std::fmt::Debug,
-    ResBody: http_body::Body,
-    ResExt: ResponseAttributeExtractor<ResBody>,
-{
-    type Output = F::Output;
-
-    fn poll(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
-        let this = self.project();
-        let _guard = this.otel_cx.clone().attach();
-        let result = std::task::ready!(this.inner.poll(cx));
-        if let Some(RequestFinalization {
-            request_data,
-            layer_state,
-            response_extractor,
-        }) = this.finalization.take()
-        {
-            finalize_request(&result, request_data, &layer_state, &response_extractor);
-        }
-        Poll::Ready(result)
-    }
-}
-
-impl<S, ReqBody, ResBody, RouteExt, ReqExt, ResExt> Service<http::Request<ReqBody>>
-    for HTTPService<S, RouteExt, ReqExt, ResExt>
-where
-    S: Service<http::Request<ReqBody>, Response = http::Response<ResBody>>,
-    S::Future: Send,
-    S::Error: std::fmt::Debug,
-    ResBody: http_body::Body,
-    RouteExt: RouteExtractor<ReqBody>,
-    ReqExt: RequestAttributeExtractor<ReqBody>,
-    ResExt: ResponseAttributeExtractor<ResBody>,
-{
-    type Response = S::Response;
-    type Error = S::Error;
-    type Future = ResponseFuture<S::Future, ResExt>;
-
-    fn poll_ready(&mut self, cx: &mut Context<'_>) -> Poll<result::Result<(), Self::Error>> {
-        self.inner_service.poll_ready(cx)
-    }
-
-    fn call(&mut self, req: http::Request<ReqBody>) -> Self::Future {
-        let duration_start = Instant::now();
-
-        let headers = req.headers();
-        let content_length = headers
-            .get(http::header::CONTENT_LENGTH)
-            .and_then(|value| value.to_str().ok()?.parse::<u64>().ok());
-
-        let (protocol, version) = split_and_format_protocol_version(req.version());
-        let protocol_name_kv = KeyValue::new(NETWORK_PROTOCOL_NAME_LABEL, protocol);
-        let protocol_version_kv = KeyValue::new(NETWORK_PROTOCOL_VERSION_LABEL, version);
-
-        // Promote the common "http"/"https" schemes to `&'static str` so the
-        // `KeyValue` is allocation-free and clones across metric label sets are cheap.
-        let url_scheme_kv = match req.uri().scheme_str() {
-            Some("http") => KeyValue::new(URL_SCHEME_LABEL, "http"),
-            Some("https") => KeyValue::new(URL_SCHEME_LABEL, "https"),
-            Some(other) => KeyValue::new(URL_SCHEME_LABEL, other.to_owned()),
-            None => KeyValue::new(URL_SCHEME_LABEL, ""),
-        };
-
-        // Same trick for well-known HTTP methods.
-        let method = req.method().as_str().to_owned();
-        let method_kv = match method_as_static(req.method()) {
-            Some(s) => KeyValue::new(HTTP_REQUEST_METHOD_LABEL, s),
-            None => KeyValue::new(HTTP_REQUEST_METHOD_LABEL, method.clone()),
-        };
-
-        // Extract route using the configured extractor
-        let route = self.route_extractor.extract_route(&req);
-        let route_kv_opt = route
-            .as_ref()
-            .map(|r| KeyValue::new(HTTP_ROUTE_LABEL, r.clone()));
-
-        // Build span name: "{method} {route}" or just "{method}"
-        let span_name = match &route {
-            Some(r) => format!("{} {}", method, r),
-            None => method.clone(),
-        };
-
-        // Extract custom request attributes
-        let custom_request_attributes = self.request_extractor.extract_attributes(&req);
-
-        // Extract the context from the incoming request headers
-        let parent_cx = global::get_text_map_propagator(|propagator| {
-            propagator.extract(&HeaderExtractor(req.headers()))
-        });
-
-        let mut span_attributes = vec![
-            KeyValue::new(semconv::trace::HTTP_REQUEST_METHOD, method.clone()),
-            url_scheme_kv.clone(),
-            KeyValue::new(semconv::attribute::URL_PATH, req.uri().path().to_string()),
-            KeyValue::new(semconv::trace::URL_FULL, req.uri().to_string()),
-        ];
-
-        if let Some(user_agent) = req
-            .headers()
-            .get("user-agent")
-            .and_then(|v| v.to_str().ok())
-        {
-            span_attributes.push(KeyValue::new(
-                semconv::trace::USER_AGENT_ORIGINAL,
-                user_agent.to_string(),
-            ));
-        }
-
-        if let Some(r) = &route {
-            span_attributes.push(KeyValue::new(HTTP_ROUTE_LABEL, r.clone()));
-        }
-
-        span_attributes.extend(custom_request_attributes.clone());
-
-        let span = self
-            .tracer
-            .span_builder(span_name)
-            .with_kind(SpanKind::Server)
-            .with_attributes(span_attributes)
-            .start_with_context(self.tracer.as_ref(), &parent_cx);
-
-        let cx = parent_cx.with_span(span);
-
-        self.state
-            .server_active_requests
-            .add(1, &[url_scheme_kv.clone(), method_kv.clone()]);
-
-        let request_data = RequestData {
-            duration_start,
-            req_body_size: content_length,
-            protocol_name_kv,
-            protocol_version_kv,
-            url_scheme_kv,
-            method_kv,
-            route_kv_opt,
-            custom_request_attributes,
-        };
-
-        let layer_state = self.state.clone();
-        let response_extractor = self.response_extractor.clone();
-
-        let inner_future = self.inner_service.call(req);
-
-        ResponseFuture {
-            inner: inner_future,
-            otel_cx: cx,
-            finalization: Some(RequestFinalization {
-                request_data,
-                layer_state,
-                response_extractor,
-            }),
-        }
-    }
-}
-
-/// Finalizes the request by updating the span and recording metrics after the response is received.
-fn finalize_request<ResBody, E, ResExt>(
-    result: &result::Result<http::Response<ResBody>, E>,
-    request_data: RequestData,
-    layer_state: &Arc<HTTPLayerState>,
-    response_extractor: &ResExt,
-) where
-    ResBody: http_body::Body,
-    ResExt: ResponseAttributeExtractor<ResBody>,
-    E: std::fmt::Debug,
-{
-    let RequestData {
-        duration_start,
-        req_body_size,
-        protocol_name_kv,
-        protocol_version_kv,
-        url_scheme_kv,
-        method_kv,
-        route_kv_opt,
-        custom_request_attributes,
-    } = request_data;
-
-    let cx = OtelContext::current();
-    let span = cx.span();
-
-    match result {
-        Ok(response) => {
-            let status = response.status();
-            let status_code_kv =
-                KeyValue::new(HTTP_RESPONSE_STATUS_CODE_LABEL, i64::from(status.as_u16()));
-
-            // Extract custom response attributes (empty/non-allocating for NoOp).
-            let custom_response_attributes = response_extractor.extract_attributes(response);
-
-            // Update span
-            span.set_attribute(status_code_kv.clone());
-            for attr in &custom_response_attributes {
-                span.set_attribute(attr.clone());
-            }
-
-            // Set span status based on HTTP status code
-            if status.is_server_error() {
-                span.set_status(Status::Error {
-                    description: format!("HTTP {}", status.as_u16()).into(),
-                });
-            }
-
-            // Build label superset by moving owned values where possible.
-            // `url_scheme_kv` and `method_kv` are cloned for the active-requests
-            // decrement; their underlying strings are typically `&'static str`
-            // so the clones are allocation-free.
-            let cap = 5
-                + route_kv_opt.is_some() as usize
-                + custom_request_attributes.len()
-                + custom_response_attributes.len();
-            let mut label_superset = Vec::with_capacity(cap);
-            label_superset.push(protocol_name_kv);
-            label_superset.push(protocol_version_kv);
-            label_superset.push(url_scheme_kv.clone());
-            label_superset.push(method_kv.clone());
-            label_superset.push(status_code_kv);
-            if let Some(route_kv) = route_kv_opt {
-                label_superset.push(route_kv);
-            }
-            // Move (not clone) the custom attribute Vecs into the label set.
-            label_superset.extend(custom_request_attributes);
-            label_superset.extend(custom_response_attributes);
-
-            layer_state
-                .server_request_duration
-                .record(duration_start.elapsed().as_secs_f64(), &label_superset);
-
-            if let Some(req_content_length) = req_body_size {
-                layer_state
-                    .server_request_body_size
-                    .record(req_content_length, &label_superset);
-            }
-
-            if let Some(resp_content_length) = response.body().size_hint().exact() {
-                layer_state
-                    .server_response_body_size
-                    .record(resp_content_length, &label_superset);
-            }
-
-            layer_state
-                .server_active_requests
-                .add(-1, &[url_scheme_kv, method_kv]);
-        }
-        Err(error) => {
-            // Mark span as error
-            span.set_status(Status::Error {
-                description: format!("{:?}", error).into(),
-            });
-
-            // Still record duration metric (without status code).
-            let label_superset = [
-                protocol_name_kv,
-                protocol_version_kv,
-                url_scheme_kv.clone(),
-                method_kv.clone(),
-            ];
-
-            layer_state
-                .server_request_duration
-                .record(duration_start.elapsed().as_secs_f64(), &label_superset);
-
-            layer_state
-                .server_active_requests
-                .add(-1, &[url_scheme_kv, method_kv]);
-        }
-    }
-}
-
-fn split_and_format_protocol_version(http_version: http::Version) -> (&'static str, &'static str) {
-    let version_str = match http_version {
-        http::Version::HTTP_09 => "0.9",
-        http::Version::HTTP_10 => "1.0",
-        http::Version::HTTP_11 => "1.1",
-        http::Version::HTTP_2 => "2.0",
-        http::Version::HTTP_3 => "3.0",
-        _ => "",
-    };
-    ("http", version_str)
-}
-
-#[cfg(test)]
-mod tests {
-    // Tests use optional provider overrides instead of global providers to avoid interference.
-    use super::*;
-
-    use http::{Request, Response, StatusCode};
-    use opentelemetry::metrics::MeterProvider;
-    use opentelemetry::trace::TracerProvider;
-    use opentelemetry::trace::{FutureExt, TraceContextExt, Tracer};
-    use opentelemetry_sdk::metrics::SdkMeterProvider;
-    use opentelemetry_sdk::metrics::{
-        data::{AggregatedMetrics, MetricData},
-        InMemoryMetricExporter, PeriodicReader,
-    };
-    use opentelemetry_sdk::trace::{InMemorySpanExporterBuilder, SdkTracerProvider};
-    use std::result::Result;
-    use std::time::Duration;
-    use tower::{Service, ServiceBuilder, ServiceExt};
-
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_tracing_with_in_memory_tracer() {
-        let trace_exporter = InMemorySpanExporterBuilder::new().build();
-        let tracer_provider = SdkTracerProvider::builder()
-            .with_simple_exporter(trace_exporter.clone())
-            .build();
-
-        let tracer = Arc::new(BoxedTracer::new(Box::new(
-            tracer_provider.tracer("test_tracer"),
-        )));
-
-        let mut layer = HTTPLayerBuilder::builder()
-            .with_route_extractor(PathExtractor)
-            .build()
-            .unwrap();
-        layer.tracer = tracer.clone();
-
-        let mut service = ServiceBuilder::new()
-            .layer(layer)
-            .service(tower::service_fn(echo));
-
-        // Create a parent span and set it as the current context
-        let parent_span = tracer.start("parent_operation");
-        let cx = OtelContext::current_with_span(parent_span);
-
-        let request_body = "test".to_string();
-        let request = http::Request::builder()
-            .uri("http://example.com/api/users/123")
-            .header("Content-Length", request_body.len().to_string())
-            .header("User-Agent", "tower-test-client/1.0")
-            .body(request_body)
-            .unwrap();
-
-        // Execute the service call within the parent span context
-        let _response = async { service.ready().await.unwrap().call(request).await.unwrap() }
-            .with_context(cx)
-            .await;
-
-        tracer_provider.force_flush().unwrap();
-
-        let spans = trace_exporter.get_finished_spans().unwrap();
-        assert_eq!(
-            spans.len(),
-            2,
-            "Expected exactly two spans to be recorded (parent + HTTP)"
-        );
-
-        // Find the HTTP span (should be the child)
-        let http_span = spans
-            .iter()
-            .find(|span| span.name == "GET /api/users/123")
-            .expect("Should find HTTP span");
-
-        // Find the parent span
-        let parent_span = spans
-            .iter()
-            .find(|span| span.name == "parent_operation")
-            .expect("Should find parent span");
-
-        // Verify the HTTP span has the correct parent
-        assert_eq!(
-            http_span.parent_span_id,
-            parent_span.span_context.span_id(),
-            "HTTP span should have parent span as parent"
-        );
-
-        // Verify they share the same trace ID
-        assert_eq!(
-            http_span.span_context.trace_id(),
-            parent_span.span_context.trace_id(),
-            "Parent and child spans should share the same trace ID"
-        );
-
-        assert_eq!(
-            http_span.name, "GET /api/users/123",
-            "Span name should match the request"
-        );
-        // Build expected attributes
-        let expected_attributes = vec![
-            KeyValue::new(semconv::trace::HTTP_REQUEST_METHOD, "GET".to_string()),
-            KeyValue::new(semconv::trace::URL_SCHEME, "http".to_string()),
-            KeyValue::new(semconv::trace::URL_PATH, "/api/users/123".to_string()),
-            KeyValue::new(
-                semconv::trace::URL_FULL,
-                "http://example.com/api/users/123".to_string(),
-            ),
-            KeyValue::new(
-                semconv::trace::USER_AGENT_ORIGINAL,
-                "tower-test-client/1.0".to_string(),
-            ),
-            KeyValue::new(semconv::trace::HTTP_ROUTE, "/api/users/123".to_string()),
-            KeyValue::new(semconv::trace::HTTP_RESPONSE_STATUS_CODE, 200),
-        ];
-
-        assert_eq!(http_span.attributes, expected_attributes);
-    }
-
-    async fn echo(req: http::Request<String>) -> Result<http::Response<String>, Error> {
-        Ok(http::Response::new(req.into_body()))
-    }
-
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_metrics_labels() {
-        let exporter = InMemoryMetricExporter::default();
-        let reader = PeriodicReader::builder(exporter.clone())
-            .with_interval(Duration::from_millis(100))
-            .build();
-        let meter_provider = SdkMeterProvider::builder().with_reader(reader).build();
-        let meter = meter_provider.meter("test");
-
-        let layer = HTTPLayerBuilder::builder()
-            .with_meter(meter)
-            .build()
-            .unwrap();
-
-        let service = tower::service_fn(|_req: Request<String>| async {
-            Ok::<_, std::convert::Infallible>(
-                Response::builder()
-                    .status(StatusCode::OK)
-                    .body(String::from("Hello, World!"))
-                    .unwrap(),
-            )
-        });
-
-        let mut service = layer.layer(service);
-
-        let request = Request::builder()
-            .method("GET")
-            .uri("https://example.com/test")
-            .body("test body".to_string())
-            .unwrap();
-
-        let _response = service.call(request).await.unwrap();
-
-        tokio::time::sleep(Duration::from_millis(500)).await;
-
-        let metrics = exporter.get_finished_metrics().unwrap();
-        assert!(!metrics.is_empty());
-
-        let resource_metrics = &metrics[0];
-        let scope_metrics = resource_metrics
-            .scope_metrics()
-            .next()
-            .expect("Should have scope metrics");
-
-        let duration_metric = scope_metrics
-            .metrics()
-            .find(|m| m.name() == HTTP_SERVER_DURATION_METRIC)
-            .expect("Duration metric should exist");
-
-        if let AggregatedMetrics::F64(MetricData::Histogram(histogram)) = duration_metric.data() {
-            let data_point = histogram
-                .data_points()
-                .next()
-                .expect("Should have data point");
-            let attributes: Vec<_> = data_point.attributes().collect();
-
-            // Duration metric should have 5 attributes: protocol_name, protocol_version, url_scheme, method, status_code
-            assert_eq!(
-                attributes.len(),
-                5,
-                "Duration metric should have exactly 5 attributes"
-            );
-
-            let protocol_name = attributes
-                .iter()
-                .find(|kv| kv.key.as_str() == NETWORK_PROTOCOL_NAME_LABEL)
-                .expect("Protocol name should be present");
-            assert_eq!(protocol_name.value.as_str(), "http");
-
-            let protocol_version = attributes
-                .iter()
-                .find(|kv| kv.key.as_str() == NETWORK_PROTOCOL_VERSION_LABEL)
-                .expect("Protocol version should be present");
-            assert_eq!(protocol_version.value.as_str(), "1.1");
-
-            let url_scheme = attributes
-                .iter()
-                .find(|kv| kv.key.as_str() == URL_SCHEME_LABEL)
-                .expect("URL scheme should be present");
-            assert_eq!(url_scheme.value.as_str(), "https");
-
-            let method = attributes
-                .iter()
-                .find(|kv| kv.key.as_str() == HTTP_REQUEST_METHOD_LABEL)
-                .expect("HTTP method should be present");
-            assert_eq!(method.value.as_str(), "GET");
-
-            let status_code = attributes
-                .iter()
-                .find(|kv| kv.key.as_str() == HTTP_RESPONSE_STATUS_CODE_LABEL)
-                .expect("Status code should be present");
-            if let opentelemetry::Value::I64(code) = &status_code.value {
-                assert_eq!(*code, 200);
-            } else {
-                panic!("Expected i64 status code");
-            }
-        } else {
-            panic!("Expected histogram data for duration metric");
-        }
-
-        let request_body_size_metric = scope_metrics
-            .metrics()
-            .find(|m| m.name() == HTTP_SERVER_REQUEST_BODY_SIZE_METRIC);
-
-        if let Some(metric) = request_body_size_metric {
-            if let AggregatedMetrics::F64(MetricData::Histogram(histogram)) = metric.data() {
-                let data_point = histogram
-                    .data_points()
-                    .next()
-                    .expect("Should have data point");
-                let attributes: Vec<_> = data_point.attributes().collect();
-
-                // Request body size metric should have 5 attributes: protocol_name, protocol_version, url_scheme, method, status_code
-                assert_eq!(
-                    attributes.len(),
-                    5,
-                    "Request body size metric should have exactly 5 attributes"
-                );
-
-                let protocol_name = attributes
-                    .iter()
-                    .find(|kv| kv.key.as_str() == NETWORK_PROTOCOL_NAME_LABEL)
-                    .expect("Protocol name should be present in request body size");
-                assert_eq!(protocol_name.value.as_str(), "https");
-
-                let protocol_version = attributes
-                    .iter()
-                    .find(|kv| kv.key.as_str() == NETWORK_PROTOCOL_VERSION_LABEL)
-                    .expect("Protocol version should be present in request body size");
-                assert_eq!(protocol_version.value.as_str(), "1.1");
-
-                let url_scheme = attributes
-                    .iter()
-                    .find(|kv| kv.key.as_str() == URL_SCHEME_LABEL)
-                    .expect("URL scheme should be present in request body size");
-                assert_eq!(url_scheme.value.as_str(), "https");
-
-                let method = attributes
-                    .iter()
-                    .find(|kv| kv.key.as_str() == HTTP_REQUEST_METHOD_LABEL)
-                    .expect("HTTP method should be present in request body size");
-                assert_eq!(method.value.as_str(), "GET");
-
-                let status_code = attributes
-                    .iter()
-                    .find(|kv| kv.key.as_str() == HTTP_RESPONSE_STATUS_CODE_LABEL)
-                    .expect("Status code should be present in request body size");
-                if let opentelemetry::Value::I64(code) = &status_code.value {
-                    assert_eq!(*code, 200);
-                } else {
-                    panic!("Expected i64 status code");
-                }
-            }
-        }
-
-        // Test response body size metric
-        let response_body_size_metric = scope_metrics
-            .metrics()
-            .find(|m| m.name() == HTTP_SERVER_RESPONSE_BODY_SIZE_METRIC);
-
-        if let Some(metric) = response_body_size_metric {
-            if let AggregatedMetrics::F64(MetricData::Histogram(histogram)) = metric.data() {
-                let data_point = histogram
-                    .data_points()
-                    .next()
-                    .expect("Should have data point");
-                let attributes: Vec<_> = data_point.attributes().collect();
-
-                // Response body size metric should have 5 attributes: protocol_name, protocol_version, url_scheme, method, status_code
-                assert_eq!(
-                    attributes.len(),
-                    5,
-                    "Response body size metric should have exactly 5 attributes"
-                );
-
-                let protocol_name = attributes
-                    .iter()
-                    .find(|kv| kv.key.as_str() == NETWORK_PROTOCOL_NAME_LABEL)
-                    .expect("Protocol name should be present in response body size");
-                assert_eq!(protocol_name.value.as_str(), "http");
-
-                let protocol_version = attributes
-                    .iter()
-                    .find(|kv| kv.key.as_str() == NETWORK_PROTOCOL_VERSION_LABEL)
-                    .expect("Protocol version should be present in response body size");
-                assert_eq!(protocol_version.value.as_str(), "1.1");
-
-                let url_scheme = attributes
-                    .iter()
-                    .find(|kv| kv.key.as_str() == URL_SCHEME_LABEL)
-                    .expect("URL scheme should be present in response body size");
-                assert_eq!(url_scheme.value.as_str(), "https");
-
-                let method = attributes
-                    .iter()
-                    .find(|kv| kv.key.as_str() == HTTP_REQUEST_METHOD_LABEL)
-                    .expect("HTTP method should be present in response body size");
-                assert_eq!(method.value.as_str(), "GET");
-
-                let status_code = attributes
-                    .iter()
-                    .find(|kv| kv.key.as_str() == HTTP_RESPONSE_STATUS_CODE_LABEL)
-                    .expect("Status code should be present in response body size");
-                if let opentelemetry::Value::I64(code) = &status_code.value {
-                    assert_eq!(*code, 200);
-                } else {
-                    panic!("Expected i64 status code");
-                }
-            }
-        }
-
-        // Test active requests metric
-        let active_requests_metric = scope_metrics
-            .metrics()
-            .find(|m| m.name() == HTTP_SERVER_ACTIVE_REQUESTS_METRIC);
-
-        if let Some(metric) = active_requests_metric {
-            if let AggregatedMetrics::I64(MetricData::Sum(sum)) = metric.data() {
-                let data_point = sum.data_points().next().expect("Should have data point");
-                let attributes: Vec<_> = data_point.attributes().collect();
-
-                // Active requests metric should have 2 attributes: method, url_scheme
-                assert_eq!(
-                    attributes.len(),
-                    2,
-                    "Active requests metric should have exactly 2 attributes"
-                );
-
-                let method = attributes
-                    .iter()
-                    .find(|kv| kv.key.as_str() == HTTP_REQUEST_METHOD_LABEL)
-                    .expect("HTTP method should be present in active requests");
-                assert_eq!(method.value.as_str(), "GET");
-
-                let url_scheme = attributes
-                    .iter()
-                    .find(|kv| kv.key.as_str() == URL_SCHEME_LABEL)
-                    .expect("URL scheme should be present in active requests");
-                assert_eq!(url_scheme.value.as_str(), "https");
-            }
-        }
-    }
-
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_context_available_in_handler() {
-        let trace_exporter = InMemorySpanExporterBuilder::new().build();
-        let tracer_provider = SdkTracerProvider::builder()
-            .with_simple_exporter(trace_exporter.clone())
-            .build();
-
-        let tracer = Arc::new(BoxedTracer::new(Box::new(
-            tracer_provider.tracer("test_tracer"),
-        )));
-
-        let mut layer = HTTPLayerBuilder::builder()
-            .with_route_extractor(PathExtractor)
-            .build()
-            .unwrap();
-        layer.tracer = tracer.clone();
-
-        let service = tower::service_fn(|_req: Request<String>| async {
-            // Access the current context - this should have the HTTP span
-            let cx = OtelContext::current();
-            let span = cx.span();
-
-            // Verify we can get span context (means context is attached)
-            let span_context = span.span_context();
-            assert!(span_context.is_valid(), "Span context should be valid");
-
-            Ok::<_, std::convert::Infallible>(
-                Response::builder()
-                    .status(StatusCode::OK)
-                    .body(String::from("OK"))
-                    .unwrap(),
-            )
-        });
-
-        let mut service = layer.layer(service);
-
-        let request = Request::builder()
-            .method("GET")
-            .uri("http://example.com/test")
-            .body("test".to_string())
-            .unwrap();
-
-        let _response = service.call(request).await.unwrap();
-
-        tracer_provider.force_flush().unwrap();
-
-        let spans = trace_exporter.get_finished_spans().unwrap();
-        assert_eq!(spans.len(), 1, "Expected one HTTP span");
-        assert_eq!(spans[0].name, "GET /test");
-    }
-
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_method_only_span_name() {
-        let trace_exporter = InMemorySpanExporterBuilder::new().build();
-        let tracer_provider = SdkTracerProvider::builder()
-            .with_simple_exporter(trace_exporter.clone())
-            .build();
-
-        let tracer = Arc::new(BoxedTracer::new(Box::new(
-            tracer_provider.tracer("test_tracer"),
-        )));
-
-        // Explicitly use method-only span names (no route extractor)
-        let mut layer = HTTPLayerBuilder::builder()
-            .with_route_extractor(NoRouteExtractor)
-            .build()
-            .unwrap();
-        layer.tracer = tracer.clone();
-
-        let service = tower::service_fn(|_req: Request<String>| async {
-            Ok::<_, std::convert::Infallible>(
-                Response::builder()
-                    .status(StatusCode::OK)
-                    .body(String::from("OK"))
-                    .unwrap(),
-            )
-        });
-
-        let mut service = layer.layer(service);
-
-        let request = Request::builder()
-            .method("POST")
-            .uri("http://example.com/users/123/orders?include=items")
-            .body("test".to_string())
-            .unwrap();
-
-        let _response = service.call(request).await.unwrap();
-
-        tracer_provider.force_flush().unwrap();
-
-        let spans = trace_exporter.get_finished_spans().unwrap();
-        assert_eq!(spans.len(), 1, "Expected one HTTP span");
-        // Should be method-only
-        assert_eq!(spans[0].name, "POST");
-    }
-
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_path_span_name_strips_query_params() {
-        let trace_exporter = InMemorySpanExporterBuilder::new().build();
-        let tracer_provider = SdkTracerProvider::builder()
-            .with_simple_exporter(trace_exporter.clone())
-            .build();
-
-        let tracer = Arc::new(BoxedTracer::new(Box::new(
-            tracer_provider.tracer("test_tracer"),
-        )));
-
-        let mut layer = HTTPLayerBuilder::builder()
-            .with_route_extractor(PathExtractor)
-            .build()
-            .unwrap();
-        layer.tracer = tracer.clone();
-
-        let service = tower::service_fn(|_req: Request<String>| async {
-            Ok::<_, std::convert::Infallible>(
-                Response::builder()
-                    .status(StatusCode::OK)
-                    .body(String::from("OK"))
-                    .unwrap(),
-            )
-        });
-
-        let mut service = layer.layer(service);
-
-        let request = Request::builder()
-            .method("GET")
-            .uri("http://example.com/users?page=1&limit=10")
-            .body("test".to_string())
-            .unwrap();
-
-        let _response = service.call(request).await.unwrap();
-
-        tracer_provider.force_flush().unwrap();
-
-        let spans = trace_exporter.get_finished_spans().unwrap();
-        assert_eq!(spans.len(), 1, "Expected one HTTP span");
-        // Query parameters should be stripped
-        assert_eq!(spans[0].name, "GET /users");
-    }
-
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_custom_span_name_extractor() {
-        let trace_exporter = InMemorySpanExporterBuilder::new().build();
-        let tracer_provider = SdkTracerProvider::builder()
-            .with_simple_exporter(trace_exporter.clone())
-            .build();
-
-        let tracer = Arc::new(BoxedTracer::new(Box::new(
-            tracer_provider.tracer("test_tracer"),
-        )));
-
-        // Custom extractor that normalizes path parameters
-        let mut layer = HTTPLayerBuilder::builder()
-            .with_route_extractor_fn(|req: &Request<String>| {
-                let path = req.uri().path();
-                // Simple normalization: replace numeric IDs with {id}
-                let normalized = path
-                    .split('/')
-                    .map(|segment| {
-                        if segment.parse::<u64>().is_ok() {
-                            "{id}"
-                        } else {
-                            segment
-                        }
-                    })
-                    .collect::<Vec<_>>()
-                    .join("/");
-                Some(normalized)
-            })
-            .build()
-            .unwrap();
-        layer.tracer = tracer.clone();
-
-        let service = tower::service_fn(|_req: Request<String>| async {
-            Ok::<_, std::convert::Infallible>(
-                Response::builder()
-                    .status(StatusCode::OK)
-                    .body(String::from("OK"))
-                    .unwrap(),
-            )
-        });
-
-        let mut service = layer.layer(service);
-
-        let request = Request::builder()
-            .method("GET")
-            .uri("http://example.com/users/12345/orders/67890")
-            .body("test".to_string())
-            .unwrap();
-
-        let _response = service.call(request).await.unwrap();
-
-        tracer_provider.force_flush().unwrap();
-
-        let spans = trace_exporter.get_finished_spans().unwrap();
-        assert_eq!(spans.len(), 1, "Expected one HTTP span");
-        // Numeric IDs should be normalized
-        assert_eq!(spans[0].name, "GET /users/{id}/orders/{id}");
-    }
-
-    #[cfg(feature = "axum")]
-    #[tokio::test(flavor = "current_thread")]
-    async fn test_axum_matched_path_extractor_fallback() {
-        let trace_exporter = InMemorySpanExporterBuilder::new().build();
-        let tracer_provider = SdkTracerProvider::builder()
-            .with_simple_exporter(trace_exporter.clone())
-            .build();
-
-        let tracer = Arc::new(BoxedTracer::new(Box::new(
-            tracer_provider.tracer("test_tracer"),
-        )));
-
-        let mut layer = HTTPLayerBuilder::builder()
-            .with_route_extractor(AxumMatchedPathExtractor)
-            .build()
-            .unwrap();
-        layer.tracer = tracer.clone();
-
-        let service = tower::service_fn(|_req: Request<String>| async {
-            Ok::<_, std::convert::Infallible>(
-                Response::builder()
-                    .status(StatusCode::OK)
-                    .body(String::from("OK"))
-                    .unwrap(),
-            )
-        });
-
-        let mut service = layer.layer(service);
-
-        // Request WITHOUT MatchedPath extension - should fall back to method-only
-        let request = Request::builder()
-            .method("GET")
-            .uri("http://example.com/users/123")
-            .body("test".to_string())
-            .unwrap();
-
-        let _response = service.call(request).await.unwrap();
-
-        tracer_provider.force_flush().unwrap();
-
-        let spans = trace_exporter.get_finished_spans().unwrap();
-        assert_eq!(spans.len(), 1, "Expected one HTTP span");
-        // Falls back to method-only when MatchedPath is not available
-        assert_eq!(spans[0].name, "GET");
     }
 }
